+++ bin/faucet/src/api/get_metadata.rs
@@ -1,8 +1,5 @@
-use std::sync::Arc;
-
 use axum::Json;
 use axum::extract::State;
-use miden_client::utils::RwLock;
 use miden_faucet_lib::FaucetId;
 use miden_faucet_lib::types::AssetAmount;
 use serde::Serialize;
@@ -17,7 +14,6 @@ use crate::api_key::ApiKey;
 #[derive(Clone)]
 pub struct Metadata {
     pub id: FaucetId,
-    pub issuance: Arc<RwLock<AssetAmount>>,
     pub max_supply: AssetAmount,
     pub decimals: u8,
     pub explorer_url: Option<Url>,
@@ -30,11 +26,9 @@ pub struct Metadata {
 #[instrument(parent = None, target = COMPONENT, name = "server.get_metadata", skip_all)]
 pub async fn get_metadata(State(server): State<ApiServer>) -> Json<GetMetadataResponse> {
     let metadata = server.metadata;
-    let issuance = metadata.issuance.read().base_units();
     Json(GetMetadataResponse {
         version: env!("CARGO_PKG_VERSION").to_string(),
         id: metadata.id.to_bech32(),
-        issuance,
         max_supply: metadata.max_supply.base_units(),
         decimals: metadata.decimals,
         explorer_url: metadata.explorer_url,
@@ -47,7 +41,6 @@ pub async fn get_metadata(State(server): State<ApiServer>) -> Json<GetMetadataRe
 pub struct GetMetadataResponse {
     pub version: String,
     pub id: String,
-    pub issuance: u64,
     pub max_supply: u64,
     pub decimals: u8,
     pub explorer_url: Option<Url>,